requirements.txt
Flask==3.1.0
Flask_WTF==1.1.1
openai==1.64.0
#protobuf==5.29.3
python-dotenv==1.0.1
Requests==2.32.3
WTForms==3.2.1
google-cloud-texttospeech==2.14.1
google-generativeai==0.3.1
Werkzeug==3.1.3
Jinja2==3.1.2